+++ src/sample.c
@@ -297,7 +297,7 @@ sample_get_graph_data(Sample *sample)
 }
 
 unsigned long
-sample_get_num_samples(Sample *sample)
+sample_get_num_sample_blocks(Sample *sample)
 {
     return sample->graph_data.numSamples;
 }